createOrder充值接口，接受double类型参数,最少0.01元（原先只接受int类型）

+++ jpress/jpress-model/src/main/java/yjt/model/base/BaseUnionpayLog.java
@@ -144,11 +144,11 @@ public class BaseUnionpayLog<M extends BaseUnionpayLog<M>> extends JModel<M> imp
 		return userId instanceof BigInteger ? (BigInteger)userId : new BigInteger(userId.toString());
 	}
 	
-	public void setFee(int fee) {
+	public void setFee(double fee) {
 		set("fee", fee);
 	}
 	
-	public int getFee() {
+	public Double getFee() {
 		return get("fee");
 	}
